+++ src/index.js
@@ -4,12 +4,11 @@ export {VERSION} from './utils/globals';
 
 export {default as Log} from './log/log';
 
-export {
-  default as Bench,
-  logResultsAsMarkdownTable,
-  logResultsAsTree,
-  logResultsAsTreeWithElapsed
-} from './bench/bench';
+// Experimental exports
+import {enableDOMLogging} from './utils/log-to-dom';
+export const experimental = {
+  enableDOMLogging
+};
 
 // // Make Probe available as global variable for debugging purposes
 // // TODO - convenient but can be a concern for "purists"
@@ -17,11 +16,3 @@ export {
 // import Probe from './probe';
 // import {global} from './utils/globals';
 // global.Probe = Probe;
-
-export {default as NodeTestDriver} from './test/node-test-driver';
-
-// Experimental exports
-import {enableDOMLogging} from './utils/log-to-dom';
-export const experimental = {
-  enableDOMLogging
-};