+++ test/bench-index.js
@@ -1,6 +1,6 @@
-import {Bench} from 'probe.gl';
+import {Bench} from 'probe.gl/bench';
 
-import colorBench from './src/color.bench';
+import colorBench from './src/bench/parse-color.bench';
 
 const suite = new Bench({
   id: 'test'