+++ test/src/test/index.js
@@ -1,0 +1,2 @@
+import './make-spy.spec';
+import './node-test-driver.spec';