+++ test/test-index.js
@@ -18,6 +18,6 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-import './src/log.spec';
-import './src/bench.spec';
-// import './src/probe.spec';
+import './src/test';
+import './src/bench';
+import './src/log';