+++ test/test-node-dist.js
@@ -18,15 +18,7 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-// Enables ES2015 import/export in Node.js
-require('reify');
-
-// Registers an alias for this module
-const path = require('path');
-const moduleAlias = require('module-alias');
-moduleAlias.addAlias('probe.gl', path.resolve('./dist'));
-
-require('babel-polyfill');
+require('./node-aliases');
 
 // Run the tests
 require('./test-index');