+++ test/test-node.js
@@ -19,14 +19,7 @@
 // THE SOFTWARE.
 
 // Enables ES2015 import/export in Node.js
-require('reify');
-
-// Registers an alias for this module
-const path = require('path');
-const moduleAlias = require('module-alias');
-moduleAlias.addAlias('probe.gl', path.resolve('./src'));
-
-require('babel-polyfill');
+require('./node-aliases');
 
 // Run the tests
 require('./test-index');